requirements.txt
datasets==3.3.0
FlagEmbedding==1.3.4
langchain==0.3.19
langchain_chroma==0.2.2
langchain_community==0.3.18
langchain_core==0.3.40
langchain_groq==0.2.4
langchain_openai==0.3.7
langgraph==0.3.2
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
ragas==0.2.13
streamlit==1.42.0
typing_extensions==4.12.2
